Move hydrogen datasets
- All datasets of hydrogen_etrago are now defined in the
  __init__.py script.
- The hydrogen bus insertion has been moved into the bus.py of
  hydrogen_etrago.
